+++ svg/src/main/java/jetbrains/jetpad/projectional/svg/SvgRoot.java
@@ -4,6 +4,6 @@ import jetbrains.jetpad.model.property.Property;
 import jetbrains.jetpad.model.property.ValueProperty;
 
 public class SvgRoot extends SvgElement {
-  public final Property<Integer> height = new ValueProperty<>(100);
-  public final Property<Integer> width = new ValueProperty<>(100);
+  public final Property<Double> height = new ValueProperty<>(100.0);
+  public final Property<Double> width = new ValueProperty<>(100.0);
 }
