+++ svg/src/main/java/jetbrains/jetpad/projectional/svg/toAwt/SvgRootMapper.java
@@ -15,16 +15,16 @@ public class SvgRootMapper extends SvgElementMapper<SvgRoot, SVGOMSVGElement> {
   protected void registerSynchronizers(SynchronizersConfiguration conf) {
     super.registerSynchronizers(conf);
 
-    conf.add(Synchronizers.forPropsOneWay(getSource().height, new WritableProperty<Integer>() {
+    conf.add(Synchronizers.forPropsOneWay(getSource().height, new WritableProperty<Double>() {
       @Override
-      public void set(Integer value) {
-        getTarget().setAttribute("height", Integer.toString(value));
+      public void set(Double value) {
+        getTarget().setAttribute("height", Double.toString(value));
       }
     }));
-    conf.add(Synchronizers.forPropsOneWay(getSource().width, new WritableProperty<Integer>() {
+    conf.add(Synchronizers.forPropsOneWay(getSource().width, new WritableProperty<Double>() {
       @Override
-      public void set(Integer value) {
-        getTarget().setAttribute("width", Integer.toString(value));
+      public void set(Double value) {
+        getTarget().setAttribute("width", Double.toString(value));
       }
     }));
   }
